add CompletableFutureMultiThreadingRepository
