+++ task8/build.gradle
@@ -14,7 +14,7 @@ android {
 
     defaultConfig {
         applicationId "by.itacademy.training.task8"
-        minSdkVersion 21
+        minSdkVersion 24
         targetSdkVersion 30
         versionCode 1
         versionName "1.0"
