+++ task8/src/main/java/by/itacademy/training/task8/model/repository/ThreadPoolExecutorMultiThreadingRepository.kt
@@ -2,9 +2,7 @@ package by.itacademy.training.task8.model.repository
 
 import by.itacademy.training.task8.entity.Contact
 import by.itacademy.training.task8.model.ContactsDao
-import java.util.concurrent.SynchronousQueue
-import java.util.concurrent.ThreadPoolExecutor
-import java.util.concurrent.TimeUnit
+import by.itacademy.training.task8.util.CustomThreadPoolExecutor
 
 class ThreadPoolExecutorMultiThreadingRepository(
     private val contactsDao: ContactsDao
@@ -31,20 +29,5 @@ class ThreadPoolExecutorMultiThreadingRepository(
     override fun delete(contact: Contact) {
         val task = { contactsDao.deleteContact(contact) }
         threadPoolExecutor.submit(task)
-    }
-}
-
-class CustomThreadPoolExecutor : ThreadPoolExecutor(
-    CORE_POOL_SIZE,
-    MAXIMUM_POOL_SIZE,
-    KEEP_ALIVE_TIME,
-    TimeUnit.SECONDS,
-    SynchronousQueue()
-) {
-    companion object {
-        private const val CORE_POOL_SIZE = 1
-        private const val MAXIMUM_POOL_SIZE = 10
-        private const val KEEP_ALIVE_TIME: Long = 20
-        private const val QUEUE_CAPACITY = 10
     }
 }
